switched console.log for dust.log
in a production environment (client or server) it is not appropriate to
log errors to the console.
helpers now log errors via the dust.log method using the "INFO" key so
that it is possible to activate logging of error messages via the log
level that has been assigned to the dust core.