Comparison helpers should still execute when their key parameter is set but resolves to undefined.
Previously, if a helper's key parameter was set, but the Dust variable to which it pointed did not exist, the helper would immediately abort. However, this breaks cases especially for {@ne} in which it would be completely valid if the key did not exist.

The helper will still fail to execute if a key parameter is not passed. However, if the parameter resolves to an undefined variable, it will now execute.

This is a non-backwards-compatible change. Closes #59.